delete strapi.ts in playwright
