+++ packages/strapi-client/src/index.ts
@@ -1,8 +1,10 @@
 export * from './strapi'
 export * from './feedback/create-feedback'
 export * from './scrapedPage/created-scraped-page'
+export * from './scrapedPage/get-or-create-scraped-page'
 export * from './scrapedPage/get-scraped-page-by-url'
 export * from './summary/create-summary'
 export * from './summary/get-all-summaries'
 export * from './summary/get-summary-id'
 export * from './summary/update-summary'
+export * from './summary/upsert-summary'
