+++ packages/strapi-client/src/summary/upsert-summary.ts
@@ -1,0 +1,24 @@
+import { createSummary, getSummaryId, updateSummary } from '..'
+import { ScrapeResultAndSummary } from '../scrapedPage/get-or-create-scraped-page'
+
+export const upsertWebPageSummary = async (
+  scrapeResultAndSummary: ScrapeResultAndSummary,
+  ScrapedWebPageId: string,
+) => {
+  const newSummary = {
+    summary: scrapeResultAndSummary.summary,
+    keywords: JSON.stringify(scrapeResultAndSummary.keywords),
+    largeLanguageModel: scrapeResultAndSummary.largeLanguageModel,
+    scraped_web_page: ScrapedWebPageId,
+  }
+
+  const webPageSummaryId = await getSummaryId(
+    scrapeResultAndSummary.largeLanguageModel,
+    scrapeResultAndSummary.url,
+    scrapeResultAndSummary.currentLanguage,
+  )
+
+  webPageSummaryId
+    ? await updateSummary(newSummary, webPageSummaryId)
+    : await createSummary(newSummary, scrapeResultAndSummary.currentLanguage)
+}
